arch/imxrt: add an entire RAM region support as a heap
i.MX.RT has three kinds of RAM, OCRAM, SRAM and SDRAM.
It is possible to locate idle stack on OCRAM and to use
an entire SDRAM region as a heap.
At that case, g_idle_topstack is not in SDRAM region.
It causes wrong size of heap in up_allocate_heap().

This commit fixes this issue.
If g_idle_topstack is not in range of PRIMARY_RAM_START and
PRIMARY_RAM_END, let's set entire region as a heap.

Without this commit, nxp_demo shows the free command with wrong value.
TASH>>free
              total       used       free    largest
Data:    1644141136      28496 1644112640 1644109552

With this,
TASH>>free
              total       used       free    largest
Data:      33554432      28496   33525936   33522848

It shows 32MB.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>