security: change the logic of getting HAL data in security API
security API which is in user space can't access kernel memory.
so if hal which is in kernel creates memory then security API can't
access it. instead of allocating memory in kernel space, security API
allocates memory and give it to hal to write result.
So inside the security handle, there are couple of memory to get data
from HAL and assign memory space when security_init is called. vice
versa it will be removed on security_deinit