iotbus: Add iotdev to get events from peripheral devices
Add iotdev to get events from peripheral device.
Now it's impossible to get detailed events from devices in kernel becuase
it is based on poll design. so adding this api iotbus in userspace can
get events from kernel devices. to do that iotdev is registered. and
it should be handled carefully because it's designed to support 1:1
relation only. so only iotbus can access it but others should not use it.