os/mm: Modify partition manager for static allocation of RAM partitions
Add a new config as CONFIG_MM_PARTITION_HEAP.
If this config is defined, then partitions are managed using heap.
Else, partitions are allocated statically. The partition for first
app is allocated at the start of the RAM and the partition for the
second app is allocated towards the end of the RAM. This approach is
suitable only for 2 apps. If more than 2 apps are required, then
CONFIG_MM_PARTITION_HEAP must be enabled.

Signed-off-by: Kishore SN <kishore.sn@samsung.com>