framework/messaging : Add parsing the received packet based on its version
If sender and receiver use different Messaging F/W version, receiver parses the received packet.
If receiver's version is higher than sender's, receiver parses the packet based on the message version.
Else if receiver's version is lower than sender's, receiver parses the packet based on its message version.

In message packet, there are version and offset which indicates the start of real data.

If someone wants to update the messaging f/w and the message header, new header data should be appended between the previous header and the real data.
And he/she should change the MSG_HEADER_SIZE and MSG_VERSION which are in messaging_internal.h.