testcase/kernel : Fix the abstime2tikcs test scenario
Prior test scenario was a meaningless beacuse The tick of
current time is not means. So I checked the tick change
after 101 second and 102 seconds. And checked the test results in
the virtual driver. I changed the name from TESTIOC_CLOCK_ABSTIME2TICKS
to TESTIOC_CLOCK_ABSTIME2TICKS_TEST and changed the location and number.