Add userspace object for application binaries
The userspace object is used to share some information from the
user space binary to the kernel. In case of loadable apps, we require
task_startup(), pthread_startup() and signal_handler() api's to be
shared from app to kernel. These api's are shared as pointers in the
userspace object.

If the userspace object is not included in the app binary, then the
kernel would use the api's from the common user binary instead of the
app binary. Then the call sequence would be kernel -> user binary -> app binary.
This is wrong since we dont want calls to go through the common user binary.

In this new implementation the userspace object will be placed at the beginning
of the app binary at an offset of 4 bytes. The first 4 bytes contain a pointer to
the app's heap object. This is followed by the userspace object and then the text
section of the app binary.

Signed-off-by: Kishore SN <kishore.sn@samsung.com>