Fix boot-up issue due to userspace object
The up_userspace.c is used by the app binaries as well as user binary.
But since the file is already build by the app build process, it is not
re-built during the user binary build. This is causing issues since the
user binary is missing some of the variables in the userspace object.
In order to fix this issue, we will delete the up_userspace.o object file
before the user binary build.

Signed-off-by: Kishore SN <kishore.sn@samsung.com>