apps/examples/timer: dedicated timer thread with sigwait()
This patch is to mitigate nested signal issues with
high-resolution timer interrupts. Nested signal situation
occurs if the time-period is smaller and next timer interrupt
occurs before the current interrupt signal is delivered.
This causes stack corruption due to multiple pending signals

This patch implements a dedicated thread that waits on sigwait()
It provides a deterministic wake-up time when the signal occurs
With this, we can achieve upto 30us timer events without suppressing
timer interrupts.

Signed-off-by: Nitish Ambastha <nitish.a@samsung.com>