framework/task_manager, apps/shell, utils: use static variables instead of globals
TASH prompt is used in tash_main.c and it is never changed.
Let's use a definition, TASH_PROMPT instead of const char *tash_prompt.

The tash_running is used to stop and it is rarelly used.
Let's make it as static and provide a function, tash_stop() to stop.

The tm_task_list, tm_pthread_list, tm_broadcast_msg in task_manager and
the stkmon_arr are used in file inside only. Let's add static.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>