iotivity: socket error fix with code 23
If DNS is not working(e.g. AP is unplugged), then Iotivity attempts to
connect to the cloud server with wrong IP address that AP give. Some
AP reply a wrong IP address even though internet is not connected. So
Iotivity stack fails to connect, but it doesn't call close().
So it caused a socket list to run out. therefore it makes socket error code 23