arch/arm/src/armv7-m: Move logic of managing g_nestlevel in exception_common
Earlier the logic of incremeting g_nestlevel and storing current_regs in
up_doirq function, but before serving this if higher priority interrupt
occurs, then there is a chances of corruption of current_regs value.

This patch moves logic of incrementing g_nestlevel and storing
current_regs value in critical area in exception_common to prevent
any kind of corruption.

Signed-off-by: Drashti Parikh <p.drashti@partner.samsung.com>