arch/stm32l4: check the validation of heap at each region
When bss or data section is too big and it exceeds first ram region,
g_idle_topstack could not be first RAM region. Then, that RAM region
should not be registered as heap.
This commit checks the location of g_idle_topstack and when it is in
range of ram region, it will be registered as first heap region.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>