lwnl: separate wi-fi control features in lwnl80211
lwnl80211 is the communication channel between kernel and user
space. so various type of network interface cards could use it for
communication. However lwnl only supports wi-fi only so it will be
difficult to extend it. so this commit distinguishes wi-fi control
part to <tinyara/net/if/wifi.h> header.
And 80211 reminds Wi-Fi. so eliminate 80211 string in lwnl.