framework/araui: Change the FPS calculate procedure.
Previously, when a FPS of certain frame belows a maximum FPS which was set through menuconfig, the compensation occurs in the next frame.
(i.e., if the maximum FPS is set as 30 and certain frame's FPS is 29, then the next frame's FPS compensated with 31 (if possible).)
Therefore, the compensated frame's FPS exceeds the maximum FPS value.

In this pull request, above problem is solved.