os/binfmt: Modify loadable app memory regions to reduce memory wastage
If CONFIG_OPTIMIZE_APP_RELOAD_TIME is enabled, then we define 3 MPU regions
for loadable apps. The existing implementation for this is to allocate one large
memory region to cover text, ro, data and heap requirements of the applications.
This large memory region is configured as RW memory in the MPU. Additionally, we
define two MPU regions for text and ro areas which will overlap the large RW region.
Allocating memory regions in this way can sometimes result in wastage of memory
due to the address alignment and region size requirements of the MPU. Specifically,
memory wastage occurs due to resizing the large RW memory region according to the
MPU requirements.

In order to overcome this problem, we modify the memory regions as follows:
- Allocate 2 separate regions for text and ro
- Allocate a seprate memory region for data + heap and configure it as RW region
- All 3 regions will follow the size and address alignment requirements of the MPU

Signed-off-by: Kishore SN <kishore.sn@samsung.com>