libc/time : Get the latest rule regardless of system time base (1970/01/01)
It requires the zoneinfo rules to show local time regardless of time
registered. But because of comparison with 0, it does not get the
latest rule which was registered before 1970. Because it needs to
find out which rules to apply based on the current time. For example
In Japan, the last rule registered was 1951. So the rule was not
applied properly. This commit fixes negative numbers to be stored.