os/mm: Modify method for accessing heap object in shared library
When we use shared common binary, the libumm module will be present in
the common binary. It will be used by all the apps to allocate and free
memory. But the mm module needs to have a pointer to the app's heap to
perform these operations. This commit contains the following changes:

- Store app id and app heap table at the start of data section of common binary
- When loading app, set the app id and populate the app heap table with heap pointer
- During each context swtich, change the value of app id
- Perform memory management using the app id as index into app heap table to obtain
heap object pointer for currently running application

Signed-off-by: Kishore S N <kishore.sn@gmail.com>