rtl8721csm: add -DARM_CORE_CM4 in makefile and remove temp definition in defconfig
If configuration is not defined in Kconfig, it disappears with menuconfig.
It means that user should not optimize the default configuration and
it is difficult to make new configuration.
The ARM_CORE_CM4 is mandatory in rtl8721csm board, let's define it
in Makefile and let's delete wrong usages in defconfig.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>