build/configs: Remove SRAM memory from user RAM regions
Currently, user ram has two regions, one in sram and one in sdram.
The sram region starts at 0x20200000 and sdram region starts at 0x80200000.
In this configuration, when any malloc is done, the memory can be allocated
by the mm either from sram or sdram depending on the requested size and available
memory. While loading applications, the space for the app sections is also
allocated using this method. However, if a scenario occurs where one section is
allocated from the sram and other section in sdram, then it is not possible
to access the symbols in one section from the other section, due to the limitation
on the max JUMP address range. This will result in failure of relocation of
some symbols and hence the app cannot run.

In order to fix this issue, we will remove the sram region and reduce the number
of regions for user memory to a single region.

Signed-off-by: Kishore S N <kishore.sn@gmail.com>