os/binfmt: Optimize common binary load time with hashmap implementation
Currently, the common binary exports its symbols with a name-address based
symbol table. In this approach, when loading the apps, its necessary to
perform many string comparisons to lookup the symbols in the common binary.
This commit removes the use of such a symbol table and replaces it with a
hashmap. The use of hashmap makes string search much faster. The load time
is reduced by upto 30% with this implementation.

A previous implementation of hashmap was found in the st_things implementation.
This is moved into the lib folder so that it can be used by all modules.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>