os/board/common/userspace: Reduce size of user binary
When app separation is enabled, most of the libraries are moved to
respective apps. In this case, the user binary must contain only the
USERSPACE object and the required API's for task_startup, pthread_startup
and singal_handler. So, in this commit, we remove all other unecessary
code from the user binary.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>