os/mm : Reduce memory requirement of memalign API
In existing implementation of memalign, we try to allocate (size + 2 * alignment)
bytes and then perform alignment within the chunk. In this case, memalign will
fail if we dont have a free node of more than (size + 2 * alignment) bytes.
In new implementation, we remove this condition. We search the list of freenodes
to find a suitable node which can allocate the required bytes with proper alignment.
The new method reduces size requirement but incurs some time overhead.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>