os/binfmt : Modfiy method of allocation ELF sections
If we allocate each elf section separately using memalign, then there
is high overhead due to the requirement of ALLOCNODE at the starting of
each allocation. To overcome this, we combine all the sections of the elf
and allocate a single chunk of memory. The sections are then arranged at
their corresponding power of 2 address within the chunk. In order to
ensure proper alignment boundaries, the sections need to be arranged in
descending order of the alignsize.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>