network/library: fix getifaddrs for Iotivity
getifaddrs is tempolary implementation to run Iotivity. it doesn't run
on TizenRT 3.0 loadable environment because it accesses to
g_netdevices which is defined in kernel layer to get a running NIC
name. So this commit adds new ioctl command(SIOCGIFNAME) to access interface name
from app layer. however this operation is not same to linux. so you'd
better to use it carefully.