arch/arm: Use new unblock function, up_recovery_unblock_task for a fault handler
This function is called by up_assert when fault happens.
A head of ready-to-run list is always changed because a fault handler
deactivates all RT threads of faulty binary before calling this function.
So if it is in interrupt context, it doesn't copy the current_regs into the OLD tcb.
If not, it calls up_fullcontextrestore to perform the user context switch.