os/mm: Remove user heap implementation
The user heap region is being used only in case of protected build scenario.
In the app separation scenario, the user heap acts as a duplicate region as
compared to the app heap. So, we will remove the user heap region and all the
implementation related to user heap from the TizenRT MM module. Going forward
we will only have kernel heap (in flat and app build) and app heap (in app build).

In case of flat build, all MM api's will operate on the kernel heap only.
When app build is enabled, the kmm api's will operate on the kernel heap, and the
kumm and malloc type api's will operate on the app heap of the currently running app.
Any calls to kumm or malloc type api's before any app is loaded will result in a
fault. Developers also need to be careful when using kumm and malloc type api's in
kernel side code, because these api's will allocate memory from app heap.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>