os: Removal of protected build feature from TizenRT
Protected build feature is not useful independently for any usage scenarios.
Since TizenRT supports flat build and app build, the user may choose app build
for having user-kernel separation. So, this commit removes protected build
as a feature from tizenRT. However, since the user-kernel separation which is
implemented under CONFIG_BUILD_PROTECTED is essential for app build, we will
internally continue to have this configuration. But, the user will not be able
to enable or disable this config independently.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>