arch/arm : Rename the function which unblocks tcb without saving old tcb
Rename the function which unblocks tcb without saving old tcb.
It moves the TCB to the ready-to-run list and switch context to the context of the task at the head of the ready to run list without saving old context.
If it is called in interrupt context, it doesn't copy the current_regs into the OLD tcb.
If not, it calls up_fullcontextrestore to perform the user context switch.