fixup! os: Fix issues after uheap removal
This commit fixes remain of commit 8878304.

When uheap is removed, then kumm and malloc type api's cannot be used
before app is loaded. They will be changed to kmm type api's. It is
also required to make sure that kumm api's are used for user tasks and
kmm api's are used for kernel tasks.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>