os/arch: Fix issue in fault handling assert address detection
The assert address is used to determine if the fault is in kernel
or common binary or app binary. A fault is either caused by calling
PANIC() in code or a hardware fault interrupt. In case of hardware
fault, the address which caused the fault is stored on the stack at
location REG_R15 by the hardware. So, the assert handler must use
REG_R15. But, if PANIC() is being called from app, then it goes through
PROXY_up_assert() before reaching the assert handler through a SWC
exception. So, in this case, the actual address where PANIC() is called
will be in REG_R14. Hence, REG_R15 is used for hardware fault and
REG_R14 is used for PANIC().

Signed-off-by: Kishore S N <kishore.sn@samsung.com>