examples/mpu: remove unnecessary config and change a type of config
1. remove CONFIG_MPUTEST_USE_PREBUILT_BINS

When loadable_apps was not added, this config was used to test two binaries.
But now, because of loadable_apps, this config is not necessary anymore.

2. change a type of CONFIG_EXAMPLES_MPU_TEST from menuconfig to config

Becasue sub-configs were removed with commit b25c5ed, it does not
need to be menuconfig. Because of this, there is no additional config
to be set, even mpu test is selected as below.

[*] MPU TestCase Example ----

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>