os/binfmt: Fix crash issue when reload optimization is disabled
When reload optimization is disabled, a single chunk of memory is
allocated for holding all elf sections in RAM. In this case, the
alloc[] array in binary_s structure will contain a single entry for
the whole binary and any access to bin->alloc[DATA] will be invalid.
So, we change all such access to bin->datastart which will hold the
start address of data section.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>