arch/armv7-m: Set MPU permission to RO-RO for read only region
Armv7m allows setting of different permissions for previleged and
user modes. So, previously we were setting RW(P)-RO(U) permission for
read only region. But armv8-m does not allow setting different permissions
based on the mode. So, we are forced to set RO-RO permission for all
read only regions. This commit changes the permission for armv7m read only
region to RO-RO, so that the behavior is uniform between armv7m and armv8m.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>