os/binfmt: Fix crash during load time due to mpu setting
We modify the mpu permission from RW-RO to RO-RO for read only
regions. In this case, kernel has RO permission only for the elf
RO section. In load_binary(), we are copying the elf data section
contents into ro region for load time optimization. But, since we
set the common binay mpu regions before this step, there is a memfault
when we try to copy into ro region. To fix this issue, this commit
moved the mpu setting for common binary after the data section backup.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>