os/mm: Fix memory leak issue due to mm_get_heap
We maintain a list of app heap objects in app_heap_list. When an
application is loaded for the first time, a new node is created and
added to this list. However, since the is_active variable in the node
is not set, it leads to a failure when searching the app_heap_list
for a heap corresponding to an address. In such case, the mm_free
operation fails to free the address, resulting in memory leak. The
issue is fixed by setting is_active to true when the app heap is being
added to the list.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>