os/mm: Fix memory leak in mm_get_heap during app update
In case of reload optimization being disabled, or in case of
app update, a new app heap node is created and added to the
app_heap_list in mm_get_heap.c. At this time, the previous
app_heap node is not freed. This results in memory leak when
reload or update app. This commit will fix this issue by
removing the old app_heap_node from the list when exit from
application (without reload).

Signed-off-by: Kishore S N <kishore.sn@samsung.com>