os/arch/common: Add error check for kheap allocate api
Kheap is created based on ram regions defined in the configuration.
The configuration provides the physical start address and size of each
region. It is possible that some ram region might contain data and stack
of the kernel. In such case, the kheap is created after the stack. But,
if the ram region is defined such that the data and stack regions extend
outside of the ram region, then it is an error. This commit checks for
such error condition and causes the device to assert on errors.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>