le_tc/kernel: Fix TC failure in tc_task_task_init
The tc_task_task_init TC is checking a global variable value which
will be modified by a different thread. This operation is being
performed without the use of a semaphore, which is resulting in
failure of TC sometimes. This commit adds semaphore checking when
using a global variable which is shared between two threads.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>
Co-authored-by: Vidisha Thapa <thapa.v@samsung.com>