fixup! arch/arm/common: Modify stack coloration size when stack overflow is enabled
Commit 74f82f3 added the coloration size with CONFIG_MPU_STACK_GUARD_SIZE always.
But currently TizenRT does not apply it for kernel thread so that there is no
stack guard at front of stack. If we color the region with CONFIG_MPU_STACK_GUARD_SIZE added,
it overwrites the coloration data out of stack region. (It corrupts the heap data which
is located at behind of stack)
This commit adds the uheap conditional and else statement.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>