os/arch/arm/src/armv8-m : New method for handing TZ Secure context
A secure context is created by TZ for each task by calling the context
allocate api. It returns a pointer to the secure context, which needs to be
stored somewhere. During every context switch from one task to next, we need
to store the context of previous task and load context of new task. This is
done by calling the corresponding load and store apis in TZ and passing the
pointer to the secure context of the task.

Previously, we were storing the secure context pointer in register R8 and also
in a global variable called tz_memory. And we had modified the exception handler
for svc exception to perform the store and load of secure context. However, this
implementation is causing several problems when protected build is used, because
in this case svc is also used to switch between user and kernel modes.

In order to overcome these problems, this commit modifies the method of storing the
secure context pointer and calling the load / store secure context API's.
- Secure context pointer will be stored in tcb of each task
- Exception handler of svc will not be modified. It will be same as exception_common
- TZ load / store api for secure context will be called in the arch specified
  code which handles the actual context switch, such as blocktask, unblocktask, etc..
- Fix the EXC_RETURN value for TZ enabled scenarios

Signed-off-by: Kishore S N <kishore.sn@samsung.com>