build/configs: Enable FPU config for amebad
The amebad rtl8721 chip has the FPU enabled in hardware. This results in
stacking of FPU registers by the HW during exception handling. If FPU config
is not enabled, it results in improper stack / unstack operations during
exception handling. This was resulting in fault due to improper PC address
during exception return when TZ is enabled, and RDP demo is executed.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>