build/configs/rtl8721csm: Fix syscall mismatch with TZ issue
When trustzone is enabled, we use 3 syscalls for secure context handling.
As a result, there is a mismatch in the syscall numbers used for the
user to kernel calls. This is result in error like below:
-----------------------------------------------------
Malloc secure memory failed [free heap size: 6040]
-----------------------------------------------------
To fix this, we need to reserve 3 more syscalls. CONFIG_SYS_RESERVED must
be changed from 8 to 11.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>