examples/testcase/network: fix network TC
this patch fix the following network TCs and wifi connectivity.

- tc_net_ioctl_p - Need to define FIONBIO as _FIOC(0x000b) to pass this TC.

when board try to join the wifi with CONFIG_NET_NETMGR then it do not get the mac address because in nm_get_netdev(), when it tries to match the g_netdev[] interfaces with  passed ifname in parameter,
it don't match. so modify CONFIG_WIFIMGR_STA_IFNAME=wl1 as CONFIG_WIFIMGR_STA_IFNAME=wlan0.

Signed-off-by: Deepak sharma <deepak.sh@samsung.com>