os/board/rtl8721csm :  Fix MPU setting error in rtk board
MPU region 0 usage depends on some conditions in code. Sometimes it is not used,
but the region number is allocated for it. So region 0 goes waste. So, we move
the region number allocation into the if condition.

Also, we were using a hardcoded value of 1 for mpu_set_nregion_board_specific().
If rtk uses more than 1 region, then this will result in overwriting the board
specific region by the app region. So, we modify the hardcoded value and instead
use the current region number setting from the rtk board for reserving the board
mpu regions.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>