os/binfmt : Allocate app elf sections from single ram region
The elf loader has a limitation that the max possible address range
for jump instruction in ARM32 format is +/- 16MB. So, we need to make
sure that the application elf sections are loaded within this range.
Escpecially, this becomes important when there is a jump instruction
from one section to another or from one binary to another.
In order to ensure this, we will use malloc_at() type of API and make
sure that all sections of all apps are loaded into the same RAM region.
This is important when the board has multiple RAM regions (like internal,
external, etc) which are distributed in the address space.

Following things are covered in this commit:
- Add new config to hold heap index in which app need to be loaded
- Use kmm_memalign_at() API for loading apps
- Ensure that a heap of the given index is actually initialized
- Ensure that the heap does not contain mutlipe regions (which might be far apart)

Signed-off-by: Kishore S N <kishore.sn@samsung.com>