security: apply rom and ram key index in seclink driver
seclink driver only supported 4 key slots before. This commit
separates the key slot area into rom which is unmodified and RAM which
can be modified while the board is on in a seclink driver.
A seclink driver manages as following invalid requests
1) invalid request (ex. set key to rom area)
2) invalid slot range (key index is invalid)
3) key slot status in RAM(set key to the slot in which a key is already
allocated)

In seclink driver, 0~63 key slots are valid. However key slot range
which is valid in each SE is different. Therefore SE driver should
check that the key is whether valid or not by itself.