configs/rtl8721csm: Enable debug configs for rtk
Enable CONFIG_DEBUG and CONFIG_DEBUG_ERROR for rtk defconfigs.
If this is not enabled, all the faults such as memfault, busfault,
etc are treated as unexpected irq and we dont get debug information
for them. These configs were previously disabled for wifi perf
issue, and since we found that wifi perf is not related to these
configs, we are re-enabling them.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>