os/kernel/binary_manager: Add NULL check in binary_manager_add_binlist
This commit is for crash issue in network_tc pthread. It is observed that
there is a busfault while running net_sendto_main(). This API creates some
pthreads. If we add some print statements in the pthread_exit, then crash is not
reproduced. It is found from analysis that the crash is happening due to
NULL pointer access in binary_manager_add_binlist. This is because, the pthread
seems to have exited before the binary_manager_add_binlist() is called, and
hence the tcb->group member is NULL.

In order to fix this issue, we perform NULL check in binary_manager_add_binlist
and if group member in tcb is null, then we do not need to add the tcb to the
app's thread list.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>