arch/arm/src/imxrt: Fix lldbg not working on imxrt
In imxrt lldbg is not working if CONFIG_DEBUG is disabled in defconfig. In other projects,
we enable the CONFIG_DEBUG in up_assert.c file to allow debug logs to be printed during
assertion. But, in imxrt, up_lowputc is put under CONFIG_DEBUG flag. So, we are not able
to get debug logs during a crash. Hence we removed the CONFIG_DEBUG condition for imxrt
up_lowputc.

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>