configs/rtl8721csm/scripts: Modify linker script to move app data to psram
When wifi app is ported to flat build, the SRAM region is overflowing due to
large size of data and bss sections. Hence, the data and bss section of app
modules will be moved to PSRAM. Since the app modules are distribulted into
many libraries, we will consider only the following libraries which are
particular to apps and contain only app related code.

libapps
libiotivity
libframework
libexternal