network/netmgr: start network stack when connection is established.
network stack was started when wifi was initialized before. So it
caused packet drop(ex. MLD) because wi-fi connection is not established. To
prevent this issue. network must be started when connection is
established. Therefore network stack is start and stop in following
cases.
1) join to an AP in STA mode => start the network stack
2) leave from an AP in STA mode => stop the network stack
3) start softAP in softAP mode => start the network stack
4) stop softAP in softAP mode => stop the network stack

And the stack size of the event thread which is created by wi-fi
driver was 1024. As event thread access network stack, it requires more
stack. So I increase it 2048. I think this approach is not
good because it has to create thread when there is wi-fi events. So it
would be better that network manager creates the event task and manage
those events.