os: Remove compression type from user binary header and other places
If CONFIG_COMPRESSED_BINARY is enabled, then all user binaries will
be compressed using the CONFIG_COMPRESSION_TYPE value. Hence, it is
not required to include the compression type in the binary header.
So, we remove the compression type value from binary header and also
any code related to setting the compression type as a separate attribute.
All decompression activity is performed based on CONFIG_COMPRESSED_BINARY
and CONFIG_COMPRESSION_TYPE configs.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>