kernel/binary_manager_load.c : Show the binary loading fail log
When fail to load binary because of invalid binary file, there was no logs if we don't enable the binary manager error log config.
In that case, there is no action after boot, so we cannot distinguish it is stucked or loaded fail.
So show the binary loading fail log always.

Signed-off-by: jeongchanKim <jc_.kim@samsung.com>